Safer destructuring.
Catch and report exceptions on object property getters, the same as Serilog does.

Implemented 'depth limiting', so that it will not stack-overflow or consume all memory and crash when destructuring large object graphs. Changed the destructure request to track and limit the depth.

Also, StructureValue.values was a mutable array, which seems like a bad tradeoff. Switched it back to an immutable list.